bootman: Move mount and list helpers
Restructure where helpers are located in source files. Make a few
helper functions public as part of this restructure to enable
sharing.
